Rename elios package to eleos
