+++ security/eleos/src/main/java/cloud/piranha/security/eleos/AuthenticationFilter.java
@@ -25,7 +25,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package cloud.piranha.security.elios;
+package cloud.piranha.security.eleos;
 
 import static cloud.piranha.webapp.api.SecurityManager.AuthenticateSource.PRE_REQUEST_CONTAINER;
 import static javax.servlet.http.HttpServletResponse.SC_FORBIDDEN;
