+++ security/jakarta/src/main/java/cloud/piranha/security/jakarta/JakartaSecurityManager.java
@@ -29,7 +29,7 @@ package cloud.piranha.security.jakarta;
 
 import static cloud.piranha.webapp.impl.DefaultAuthenticatedIdentity.getCurrentSubject;
 import static cloud.piranha.webapp.api.SecurityManager.AuthenticateSource.MID_REQUEST_USER;
-import static cloud.piranha.security.elios.AuthenticationInitializer.AUTH_SERVICE;
+import static cloud.piranha.security.eleos.AuthenticationInitializer.AUTH_SERVICE;
 import static cloud.piranha.security.exousia.AuthorizationPreInitializer.AUTHZ_SERVICE;
 
 import java.io.IOException;
